backend/pyproject.toml
[tool.poetry]
name = "genai-app-backend"
version = "0.1.0"
description = "Gen AI Backed for Learning"
authors = ["Rohit Tiwari <rohittiwarirvt@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11.3"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
pep8 = "^1.7.1"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.3.4"
python-dotenv = {extras = ["cli"], version = "^1.0.1"}
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
llama-index-readers-elasticsearch = "^0.1.4"
llama-index-vector-stores-opensearch = "^0.1.12"
llama-index-embeddings-ollama = "^0.1.2"
pathlib = "^1.0.1"
llama-index-readers-file = "^0.1.30"
llama-index-embeddings-openai = "^0.1.10"
llama-index-llms-openai = "^0.1.26"


uuid = "^1.30"
logging = "^0.4.9.6"
faker = "^26.0.0"
sse-starlette = "^2.1.2"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "app.restserver:start"